Refactors and moves buildConfigs and imageStreams related code into new files.
It also refactors a few other resources like deployment etc.
