+++ pkg/occlient/occlient_test.go
@@ -8,7 +8,6 @@ import (
 	"sort"
 	"strings"
 	"testing"
-	"time"
 
 	scv1beta1 "github.com/kubernetes-sigs/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/kylelemons/godebug/pretty"
@@ -25,7 +24,6 @@ import (
 	"github.com/pkg/errors"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/fields"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/watch"
 	ktesting "k8s.io/client-go/testing"
@@ -434,235 +432,6 @@ func TestParseImageName(t *testing.T) {
 	}
 }
 
-func TestUpdateDCAnnotations(t *testing.T) {
-	tests := []struct {
-		name        string
-		dcName      string
-		annotations map[string]string
-		existingDc  appsv1.DeploymentConfig
-		wantErr     bool
-	}{
-		{
-			name:   "existing dc",
-			dcName: "nodejs",
-			annotations: map[string]string{
-				"app.kubernetes.io/component-source-type": "local",
-			},
-			existingDc: appsv1.DeploymentConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs",
-					Annotations: map[string]string{"app.openshift.io/vcs-uri": "https://github.com/sclorg/nodejs-ex",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:   "non existing dc",
-			dcName: "nodejs",
-			annotations: map[string]string{
-				"app.kubernetes.io/component-source-type": "local",
-			},
-			existingDc: appsv1.DeploymentConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "wildfly",
-					Annotations: map[string]string{"app.openshift.io/vcs-uri": "https://github.com/sclorg/nodejs-ex",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-			},
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-			fkclientset.AppsClientset.PrependReactor("get", "deploymentconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				dcName := action.(ktesting.GetAction).GetName()
-				if dcName != tt.dcName {
-					return true, nil, fmt.Errorf("'get' called with a different dcName")
-				}
-
-				if tt.dcName != tt.existingDc.Name {
-					return true, nil, fmt.Errorf("got different dc")
-				}
-				return true, &tt.existingDc, nil
-			})
-
-			fkclientset.AppsClientset.PrependReactor("update", "deploymentconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				dc := action.(ktesting.UpdateAction).GetObject().(*appsv1.DeploymentConfig)
-				if dc.Name != tt.existingDc.Name {
-					return true, nil, fmt.Errorf("got different dc")
-				}
-				return true, dc, nil
-			})
-
-			err := fkclient.UpdateDCAnnotations(tt.dcName, tt.annotations)
-
-			if err == nil && !tt.wantErr {
-				// Check for validating actions performed
-				if (len(fkclientset.AppsClientset.Actions()) != 2) && (tt.wantErr != true) {
-					t.Errorf("expected 2 action in UpdateDeploymentConfig got: %v", fkclientset.AppsClientset.Actions())
-				}
-
-				updatedDc := fkclientset.AppsClientset.Actions()[1].(ktesting.UpdateAction).GetObject().(*appsv1.DeploymentConfig)
-				if updatedDc.Name != tt.dcName {
-					t.Errorf("deploymentconfig name is not matching with expected value, expected: %s, got %s", tt.dcName, updatedDc.Name)
-				}
-
-				if !reflect.DeepEqual(updatedDc.Annotations, tt.annotations) {
-					t.Errorf("deployment Config annotations not matching with expected values, expected: %s, got %s", tt.annotations, updatedDc.Annotations)
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
-func TestGetBuildConfigFromName(t *testing.T) {
-	tests := []struct {
-		name                string
-		buildName           string
-		returnedBuildConfig buildv1.BuildConfig
-		wantErr             bool
-	}{
-		{
-			name:      "buildConfig with existing bc",
-			buildName: "nodejs",
-			returnedBuildConfig: buildv1.BuildConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs",
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-
-			fkclientset.BuildClientset.PrependReactor("get", "buildconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				buildName := action.(ktesting.GetAction).GetName()
-				if buildName != tt.buildName {
-					return true, nil, fmt.Errorf("'get' was called with wrong buildName")
-				}
-				return true, &tt.returnedBuildConfig, nil
-			})
-
-			build, err := fkclient.GetBuildConfigFromName(tt.buildName)
-			if err == nil && !tt.wantErr {
-				// Check for validating actions performed
-				if (len(fkclientset.BuildClientset.Actions()) != 1) && (tt.wantErr != true) {
-					t.Errorf("expected 1 action in GetBuildConfigFromName got: %v", fkclientset.AppsClientset.Actions())
-				}
-				if build.Name != tt.buildName {
-					t.Errorf("wrong GetBuildConfigFromName got: %v, expected: %v", build.Name, tt.buildName)
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
-func TestUpdateBuildConfig(t *testing.T) {
-	tests := []struct {
-		name                string
-		buildConfigName     string
-		gitURL              string
-		annotations         map[string]string
-		existingBuildConfig buildv1.BuildConfig
-		updatedBuildConfig  buildv1.BuildConfig
-		wantErr             bool
-	}{
-		{
-			name:            "local to git with proper parameters",
-			buildConfigName: "nodejs",
-			gitURL:          "https://github.com/sclorg/nodejs-ex",
-			annotations: map[string]string{
-				"app.openshift.io/vcs-uri":                "https://github.com/sclorg/nodejs-ex",
-				"app.kubernetes.io/component-source-type": "git",
-			},
-			existingBuildConfig: buildv1.BuildConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs",
-				},
-				Spec: buildv1.BuildConfigSpec{
-					CommonSpec: buildv1.CommonSpec{},
-				},
-			},
-			updatedBuildConfig: buildv1.BuildConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs",
-					Annotations: map[string]string{
-						"app.openshift.io/vcs-uri":                "https://github.com/sclorg/nodejs-ex",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-				Spec: buildv1.BuildConfigSpec{
-					CommonSpec: buildv1.CommonSpec{
-						Source: buildv1.BuildSource{
-							Git: &buildv1.GitBuildSource{
-								URI: "https://github.com/sclorg/nodejs-ex",
-							},
-							Type: buildv1.BuildSourceGit,
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-			fkclientset.BuildClientset.PrependReactor("get", "buildconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				buildConfigName := action.(ktesting.GetAction).GetName()
-				if buildConfigName != tt.buildConfigName {
-					return true, nil, fmt.Errorf("'update' was called with wrong buildConfig name")
-				}
-				return true, &tt.existingBuildConfig, nil
-			})
-
-			fkclientset.BuildClientset.PrependReactor("update", "buildconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				buildConfig := action.(ktesting.UpdateAction).GetObject().(*buildv1.BuildConfig)
-				if buildConfig.Name != tt.buildConfigName {
-					return true, nil, fmt.Errorf("'update' was called with wrong buildConfig name")
-				}
-				return true, &tt.updatedBuildConfig, nil
-			})
-
-			err := fkclient.UpdateBuildConfig(tt.buildConfigName, tt.gitURL, tt.annotations)
-			if err == nil && !tt.wantErr {
-				// Check for validating actions performed
-				if (len(fkclientset.BuildClientset.Actions()) != 2) && (tt.wantErr != true) {
-					t.Errorf("expected 2 action in GetBuildConfigFromName got: %v", fkclientset.BuildClientset.Actions())
-				}
-
-				updatedDc := fkclientset.BuildClientset.Actions()[1].(ktesting.UpdateAction).GetObject().(*buildv1.BuildConfig)
-				if !reflect.DeepEqual(updatedDc.Annotations, tt.annotations) {
-					t.Errorf("deployment Config annotations not matching with expected values, expected: %s, got %s", tt.annotations, updatedDc.Annotations)
-				}
-
-				if !reflect.DeepEqual(updatedDc.Spec, tt.updatedBuildConfig.Spec) {
-					t.Errorf("deployment Config Spec not matching with expected values: %v", pretty.Compare(tt.updatedBuildConfig.Spec, updatedDc.Spec))
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
 func TestNewAppS2I(t *testing.T) {
 	type args struct {
 		commonObjectMeta   metav1.ObjectMeta
@@ -1003,7 +772,7 @@ func TestIsTagInImageStream(t *testing.T) {
 	}
 }
 
-func TestGetExposedPorts(t *testing.T) {
+func Test_getExposedPortsFromISI(t *testing.T) {
 	tests := []struct {
 		name             string
 		imageTag         string
@@ -1129,7 +898,7 @@ func TestGetExposedPorts(t *testing.T) {
 		t.Run(tt.name, func(t *testing.T) {
 			fkclient, _ := FakeNew()
 			fkclient.Namespace = "testing"
-			got, err := fkclient.GetExposedPorts(tt.imageStreamImage)
+			got, err := getExposedPortsFromISI(tt.imageStreamImage)
 
 			// sort result, map is used behind the scene so the ordering might be different
 			sort.Slice(got, func(i, j int) bool {
@@ -1488,320 +1257,6 @@ func TestUnlinkSecret(t *testing.T) {
 			}
 		})
 	}
-}
-
-func TestGetImageStream(t *testing.T) {
-	tests := []struct {
-		name           string
-		imageNS        string
-		imageName      string
-		imageTag       string
-		wantErr        bool
-		want           *imagev1.ImageStream
-		wantActionsCnt int
-	}{
-		{
-			name:           "Case: Valid request for imagestream of latest version and not namespace qualified",
-			imageNS:        "",
-			imageName:      "foo",
-			imageTag:       "latest",
-			want:           fakeImageStream("foo", "testing", []string{"latest"}),
-			wantActionsCnt: 1,
-		},
-		{
-			name:           "Case: Valid explicit request for specific namespace qualified imagestream of specific version",
-			imageNS:        "openshift",
-			imageName:      "foo",
-			imageTag:       "latest",
-			want:           fakeImageStream("foo", "openshift", []string{"latest", "3.5"}),
-			wantActionsCnt: 1,
-		},
-		{
-			name:           "Case: Valid request for specific imagestream of specific version not in current namespace",
-			imageNS:        "",
-			imageName:      "foo",
-			imageTag:       "3.5",
-			want:           fakeImageStream("foo", "openshift", []string{"latest", "3.5"}),
-			wantActionsCnt: 1, // Ideally supposed to be 2 but bcoz prependreactor is not parameter sensitive, the way it is mocked makes it 1
-		},
-		{
-			name:           "Case: Invalid request for non-current and non-openshift namespace imagestream/Non-existant imagestream",
-			imageNS:        "foo",
-			imageName:      "bar",
-			imageTag:       "3.5",
-			wantErr:        true,
-			wantActionsCnt: 1,
-		},
-		{
-			name:           "Case: Request for non-existant tag",
-			imageNS:        "",
-			imageName:      "foo",
-			imageTag:       "3.6",
-			wantErr:        true,
-			wantActionsCnt: 2,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-			fkclient.Namespace = "testing"
-			openshiftIS := fakeImageStream(tt.imageName, "openshift", []string{"latest", "3.5"})
-			currentNSIS := fakeImageStream(tt.imageName, "testing", []string{"latest"})
-
-			fkclientset.ImageClientset.PrependReactor("get", "imagestreams", func(action ktesting.Action) (bool, runtime.Object, error) {
-				if tt.imageNS == "" {
-					if isTagInImageStream(*fakeImageStream("foo", "testing", []string{"latest"}), tt.imageTag) {
-						return true, currentNSIS, nil
-					} else if isTagInImageStream(*fakeImageStream("foo", "openshift", []string{"latest", "3.5"}), tt.imageTag) {
-						return true, openshiftIS, nil
-					}
-					return true, nil, fmt.Errorf("Requested imagestream %s with tag %s not found", tt.imageName, tt.imageTag)
-				}
-				if tt.imageNS == "testing" {
-					return true, currentNSIS, nil
-				}
-				if tt.imageNS == "openshift" {
-					return true, openshiftIS, nil
-				}
-				return true, nil, fmt.Errorf("Requested imagestream %s with tag %s not found", tt.imageName, tt.imageTag)
-			})
-
-			got, err := fkclient.GetImageStream(tt.imageNS, tt.imageName, tt.imageTag)
-			if len(fkclientset.ImageClientset.Actions()) != tt.wantActionsCnt {
-				t.Errorf("expected %d ImageClientset.Actions() in GetImageStream, got %v", tt.wantActionsCnt, fkclientset.ImageClientset.Actions())
-			}
-			if !tt.wantErr == (err != nil) {
-				t.Errorf("\nclient.GetImageStream(imageNS, imageName, imageTag) unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("GetImageStream() = %#v, want %#v and the current project name is %s\n\n", got, tt, fkclient.GetCurrentProjectName())
-			}
-		})
-	}
-}
-
-func TestGetImageStreams(t *testing.T) {
-
-	type args struct {
-		name      string
-		namespace string
-	}
-
-	tests := []struct {
-		name    string
-		args    args
-		want    []imagev1.ImageStream
-		wantErr bool
-	}{
-		{
-			name: "case 1: testing a valid imagestream",
-			args: args{
-				name:      "ruby",
-				namespace: "testing",
-			},
-			want: []imagev1.ImageStream{
-				{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "ruby",
-						Namespace: "testing",
-					},
-					Status: imagev1.ImageStreamStatus{
-						Tags: []imagev1.NamedTagEventList{
-							{
-								Tag: "latest",
-								Items: []imagev1.TagEvent{
-									{
-										DockerImageReference: "example/ruby:latest",
-										Generation:           1,
-										Image:                "sha256:9579a93ee",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			wantErr: false,
-		},
-
-		// TODO: Currently fails. Enable once fixed
-		// {
-		//         name: "case 2: empty namespace",
-		//         args: args{
-		//                 name:      "ruby",
-		//                 namespace: "",
-		//         },
-		//         wantErr: true,
-		// },
-
-		// {
-		// 	name: "case 3: empty name",
-		// 	args: args{
-		// 		name:      "",
-		// 		namespace: "testing",
-		// 	},
-		// 	wantErr: true,
-		// },
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			client, fkclientset := FakeNew()
-
-			fkclientset.ImageClientset.PrependReactor("list", "imagestreams", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, fakeImageStreams(tt.args.name, tt.args.namespace), nil
-			})
-
-			got, err := client.GetImageStreams(tt.args.namespace)
-
-			if (err != nil) != tt.wantErr {
-				t.Errorf("GetImageStreams() error = %#v, wantErr %#v", err, tt.wantErr)
-				return
-			}
-
-			if len(fkclientset.ImageClientset.Actions()) != 1 {
-				t.Errorf("expected 1 action in GetImageStreams got: %v", fkclientset.ImageClientset.Actions())
-			}
-
-			if !reflect.DeepEqual(got, tt.want) {
-				t.Errorf("GetImageStreams() = %#v, want %#v", got, tt.want)
-			}
-
-		})
-	}
-}
-
-func TestStartBuild(t *testing.T) {
-	tests := []struct {
-		name    string
-		bcName  string
-		wantErr bool
-	}{
-		{
-			name:    "Case 1: Testing valid name",
-			bcName:  "ruby",
-			wantErr: false,
-		},
-
-		// TODO: Currently fails. Enable once fixed.
-		// {
-		// 	name:    "Case 2: Testing empty name",
-		// 	bcName:  "",
-		// 	wantErr: true,
-		// },
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fkclient, fkclientset := FakeNew()
-
-			fkclientset.BuildClientset.PrependReactor("create", "buildconfigs", func(action ktesting.Action) (bool, runtime.Object, error) {
-				build := buildv1.Build{
-					ObjectMeta: metav1.ObjectMeta{
-						Name: tt.bcName,
-					},
-				}
-
-				return true, &build, nil
-			})
-
-			_, err := fkclient.StartBuild(tt.bcName)
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.StartBuild(string) unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-
-			if err == nil {
-
-				if len(fkclientset.BuildClientset.Actions()) != 1 {
-					t.Errorf("expected 1 action in StartBuild got: %v", fkclientset.BuildClientset.Actions())
-				}
-
-				startedBuild := fkclientset.BuildClientset.Actions()[0].(ktesting.CreateAction).GetObject().(*buildv1.BuildRequest)
-
-				if startedBuild.Name != tt.bcName {
-					t.Errorf("buildconfig name is not matching to expected name, expected: %s, got %s", tt.bcName, startedBuild.Name)
-				}
-			}
-		})
-	}
-
-}
-
-func TestWaitForBuildToFinish(t *testing.T) {
-
-	tests := []struct {
-		name      string
-		buildName string
-		status    buildv1.BuildPhase
-		wantErr   bool
-	}{
-		{
-			name:      "phase: complete",
-			buildName: "ruby",
-			status:    buildv1.BuildPhaseComplete,
-			wantErr:   false,
-		},
-
-		{
-			name:      "phase: failed",
-			buildName: "ruby",
-			status:    buildv1.BuildPhaseFailed,
-			wantErr:   true,
-		},
-
-		{
-			name:      "phase: cancelled",
-			buildName: "ruby",
-			status:    buildv1.BuildPhaseCancelled,
-			wantErr:   true,
-		},
-
-		{
-			name:      "phase: error",
-			buildName: "ruby",
-			status:    buildv1.BuildPhaseError,
-			wantErr:   true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-
-			fkclient, fkclientset := FakeNew()
-			fkWatch := watch.NewFake()
-
-			go func() {
-				fkWatch.Modify(fakeBuildStatus(tt.status, tt.buildName))
-			}()
-
-			fkclientset.BuildClientset.PrependWatchReactor("builds", func(action ktesting.Action) (handled bool, ret watch.Interface, err error) {
-				return true, fkWatch, nil
-			})
-
-			err := fkclient.WaitForBuildToFinish(tt.buildName, os.Stdout, time.Second*5)
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" client.WaitForBuildToFinish(string) unexpected error %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-
-			if len(fkclientset.BuildClientset.Actions()) != 1 {
-				t.Errorf("expected 1 action in WaitForBuildToFinish got: %v", fkclientset.BuildClientset.Actions())
-			}
-
-			if err == nil {
-				expectedFields := fields.OneTermEqualSelector("metadata.name", tt.buildName)
-				gotFields := fkclientset.BuildClientset.Actions()[0].(ktesting.WatchAction).GetWatchRestrictions().Fields
-
-				if !reflect.DeepEqual(expectedFields, gotFields) {
-					t.Errorf("Fields not matching: expected: %s, got %s", expectedFields, gotFields)
-				}
-			}
-		})
-	}
-
 }
 
 func TestCreateServiceInstance(t *testing.T) {
@@ -3146,774 +2601,6 @@ func TestInjectS2IPaths(t *testing.T) {
 	}
 }
 
-func TestGetS2IMetaInfoFromBuilderImg(t *testing.T) {
-	tests := []struct {
-		name             string
-		imageStreamImage *imagev1.ImageStreamImage
-		want             S2IPaths
-		wantErr          bool
-	}{
-		{
-			name: "Case 1: Valid nodejs test case with image protocol access",
-			imageStreamImage: fakeImageStreamImage(
-				"nodejs",
-				[]string{"8080/tcp"},
-				`{
-					"kind": "DockerImage",
-					"apiVersion": "1.0",
-					"Id": "sha256:93de1230c12b512ebbaf28b159f450a44c632eda06bdc0754236f403f5876234",
-					"Created": "2018-10-19T15:43:13Z",
-					"ContainerConfig": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"#(nop) ",
-							"CMD [\"/bin/sh\" \"-c\" \"${STI_SCRIPTS_PATH}/usage\"]"
-						],
-						"Image": "sha256:d353b3f467c2d2ff59a1e09bb91cff1c493aedd6c8041ebb273346e892f8ee85",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"DockerVersion": "18.06.0-ce",
-					"Config": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"${STI_SCRIPTS_PATH}/usage"
-						],
-						"Image": "57a00ab03a3f8c3af19e91c284e0c499b16d5115c925aa845b5d14eace949c34",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"Architecture": "amd64",
-					"Size": 221580439
-				}`,
-			),
-			want: S2IPaths{
-				ScriptsPathProtocol: "image://",
-				ScriptsPath:         "/usr/libexec/s2i",
-				SrcOrBinPath:        "/tmp",
-				DeploymentDir:       "/opt/app-root/src",
-				WorkingDir:          "/opt/app-root/src",
-				SrcBackupPath:       "/opt/app-root/src-backup",
-				BuilderImgName:      "bucharestgold/centos7-s2i-nodejs",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Case 2: Valid nodejs test case with file protocol access",
-			imageStreamImage: fakeImageStreamImage(
-				"nodejs",
-				[]string{"8080/tcp"},
-				`{
-					"kind": "DockerImage",
-					"apiVersion": "1.0",
-					"Id": "sha256:93de1230c12b512ebbaf28b159f450a44c632eda06bdc0754236f403f5876234",
-					"Created": "2018-10-19T15:43:13Z",
-					"ContainerConfig": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=file:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"#(nop) ",
-							"CMD [\"/bin/sh\" \"-c\" \"${STI_SCRIPTS_PATH}/usage\"]"
-						],
-						"Image": "sha256:d353b3f467c2d2ff59a1e09bb91cff1c493aedd6c8041ebb273346e892f8ee85",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "file:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "file:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"DockerVersion": "18.06.0-ce",
-					"Config": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"${STI_SCRIPTS_PATH}/usage"
-						],
-						"Image": "57a00ab03a3f8c3af19e91c284e0c499b16d5115c925aa845b5d14eace949c34",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "file:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"Architecture": "amd64",
-					"Size": 221580439
-				}`,
-			),
-			want: S2IPaths{
-				ScriptsPathProtocol: "file://",
-				ScriptsPath:         "/usr/libexec/s2i",
-				SrcOrBinPath:        "/tmp",
-				DeploymentDir:       "/opt/app-root/src",
-				WorkingDir:          "/opt/app-root/src",
-				SrcBackupPath:       "/opt/app-root/src-backup",
-				BuilderImgName:      "bucharestgold/centos7-s2i-nodejs",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Case 3: Valid nodejs test case with http(s) protocol access",
-			imageStreamImage: fakeImageStreamImage(
-				"nodejs",
-				[]string{"8080/tcp"},
-				`{
-					"kind": "DockerImage",
-					"apiVersion": "1.0",
-					"Id": "sha256:93de1230c12b512ebbaf28b159f450a44c632eda06bdc0754236f403f5876234",
-					"Created": "2018-10-19T15:43:13Z",
-					"ContainerConfig": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=http(s):///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"#(nop) ",
-							"CMD [\"/bin/sh\" \"-c\" \"${STI_SCRIPTS_PATH}/usage\"]"
-						],
-						"Image": "sha256:d353b3f467c2d2ff59a1e09bb91cff1c493aedd6c8041ebb273346e892f8ee85",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "http(s):///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"DockerVersion": "18.06.0-ce",
-					"Config": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"${STI_SCRIPTS_PATH}/usage"
-						],
-						"Image": "57a00ab03a3f8c3af19e91c284e0c499b16d5115c925aa845b5d14eace949c34",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "http(s):///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"Architecture": "amd64",
-					"Size": 221580439
-				}`,
-			),
-			want: S2IPaths{
-				ScriptsPathProtocol: "http(s)://",
-				ScriptsPath:         "http(s):///usr/libexec/s2i",
-				SrcOrBinPath:        "/tmp",
-				DeploymentDir:       "/opt/app-root/src",
-				WorkingDir:          "/opt/app-root/src",
-				SrcBackupPath:       "/opt/app-root/src-backup",
-				BuilderImgName:      "bucharestgold/centos7-s2i-nodejs",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Case 4: Valid openjdk test case with image(s) protocol access",
-			imageStreamImage: fakeImageStreamImage(
-				"redhat-openjdk18-openshift",
-				[]string{"8080/tcp"},
-				`{
-					"kind": "DockerImage",
-					"apiVersion": "1.0",
-					"Id": "sha256:93de1230c12b512ebbaf28b159f450a44c632eda06bdc0754236f403f5876234",
-					"Created": "2018-10-19T15:43:13Z",
-					"ContainerConfig": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=http(s):///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"#(nop) ",
-							"CMD [\"/bin/sh\" \"-c\" \"${STI_SCRIPTS_PATH}/usage\"]"
-						],
-						"Image": "sha256:d353b3f467c2d2ff59a1e09bb91cff1c493aedd6c8041ebb273346e892f8ee85",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"DockerVersion": "18.06.0-ce",
-					"Config": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"${STI_SCRIPTS_PATH}/usage"
-						],
-						"Image": "57a00ab03a3f8c3af19e91c284e0c499b16d5115c925aa845b5d14eace949c34",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"Architecture": "amd64",
-					"Size": 221580439
-				}`,
-			),
-			want: S2IPaths{
-				ScriptsPathProtocol: "image://",
-				ScriptsPath:         "/usr/libexec/s2i",
-				SrcOrBinPath:        "/tmp",
-				DeploymentDir:       "/opt/app-root/src",
-				WorkingDir:          "/opt/app-root/src",
-				SrcBackupPath:       "/opt/app-root/src-backup",
-				BuilderImgName:      "bucharestgold/centos7-s2i-nodejs",
-			},
-			wantErr: false,
-		},
-		{
-			name: "Case 5: Inalid nodejs test case with invalid protocol access",
-			imageStreamImage: fakeImageStreamImage(
-				"nodejs",
-				[]string{"8080/tcp"},
-				`{
-					"kind": "DockerImage",
-					"apiVersion": "1.0",
-					"Id": "sha256:93de1230c12b512ebbaf28b159f450a44c632eda06bdc0754236f403f5876234",
-					"Created": "2018-10-19T15:43:13Z",
-					"ContainerConfig": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=something:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"#(nop) ",
-							"CMD [\"/bin/sh\" \"-c\" \"${STI_SCRIPTS_PATH}/usage\"]"
-						],
-						"Image": "sha256:d353b3f467c2d2ff59a1e09bb91cff1c493aedd6c8041ebb273346e892f8ee85",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "something:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"DockerVersion": "18.06.0-ce",
-					"Config": {
-						"Hostname": "8911994b686d",
-						"User": "1001",
-						"ExposedPorts": {
-							"8080/tcp": {}
-						},
-						"Env": [
-							"PATH=/opt/app-root/src/node_modules/.bin/:/opt/app-root/src/.npm-global/bin/:/opt/app-root/src/bin:/opt/app-root/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin",
-							"SUMMARY=Platform for building and running Node.js 10.12.0 applications",
-							"DESCRIPTION=Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"STI_SCRIPTS_URL=image:///usr/libexec/s2i",
-							"STI_SCRIPTS_PATH=/usr/libexec/s2i",
-							"APP_ROOT=/opt/app-root",
-							"HOME=/opt/app-root/src",
-							"BASH_ENV=/opt/app-root/etc/scl_enable",
-							"ENV=/opt/app-root/etc/scl_enable",
-							"PROMPT_COMMAND=. /opt/app-root/etc/scl_enable",
-							"NODEJS_SCL=rh-nodejs8",
-							"NPM_RUN=start",
-							"NODE_VERSION=10.12.0",
-							"NPM_VERSION=6.4.1",
-							"NODE_LTS=false",
-							"NPM_CONFIG_LOGLEVEL=info",
-							"NPM_CONFIG_PREFIX=/opt/app-root/src/.npm-global",
-							"NPM_CONFIG_TARBALL=/usr/share/node/node-v10.12.0-headers.tar.gz",
-							"DEBUG_PORT=5858"
-						],
-						"Cmd": [
-							"/bin/sh",
-							"-c",
-							"${STI_SCRIPTS_PATH}/usage"
-						],
-						"Image": "57a00ab03a3f8c3af19e91c284e0c499b16d5115c925aa845b5d14eace949c34",
-						"WorkingDir": "/opt/app-root/src",
-						"Entrypoint": [
-							"container-entrypoint"
-						],
-						"Labels": {
-							"com.redhat.component": "s2i-base-container",
-							"com.redhat.deployments-dir": "/opt/app-root/src",
-							"com.redhat.dev-mode": "DEV_MODE:false",
-							"com.rehdat.dev-mode.port": "DEBUG_PORT:5858",
-							"description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.description": "Node.js  available as docker container is a base platform for building and running various Node.js  applications and frameworks. Node.js is a platform built on Chrome's JavaScript runtime for easily building fast, scalable network applications. Node.js uses an event-driven, non-blocking I/O model that makes it lightweight and efficient, perfect for data-intensive real-time applications that run across distributed devices.",
-							"io.k8s.display-name": "Node.js 10.12.0",
-							"io.openshift.builder-version": "\"190ef14\"",
-							"io.openshift.expose-services": "8080:http",
-							"io.openshift.s2i.scripts-url": "something:///usr/libexec/s2i",
-							"io.openshift.tags": "builder,nodejs,nodejs-10.12.0",
-							"io.s2i.scripts-url": "image:///usr/libexec/s2i",
-							"maintainer": "Lance Ball \u003clball@redhat.com\u003e",
-							"name": "bucharestgold/centos7-s2i-nodejs",
-							"org.label-schema.build-date": "20180804",
-							"org.label-schema.license": "GPLv2",
-							"org.label-schema.name": "CentOS Base Image",
-							"org.label-schema.schema-version": "1.0",
-							"org.label-schema.vendor": "CentOS",
-							"release": "1",
-							"summary": "Platform for building and running Node.js 10.12.0 applications",
-							"version": "10.12.0"
-						}
-					},
-					"Architecture": "amd64",
-					"Size": 221580439
-				}`,
-			),
-			want:    S2IPaths{},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			s2iPaths, err := GetS2IMetaInfoFromBuilderImg(tt.imageStreamImage)
-			if !reflect.DeepEqual(tt.want, s2iPaths) {
-				t.Errorf("s2i paths are not matching with expected values, expected: %v, got %v", tt.want, s2iPaths)
-			}
-			if !tt.wantErr == (err != nil) {
-				t.Errorf(" GetS2IScriptsPathFromBuilderImg() unexpected error %v, wantErr %v", err, tt.wantErr)
-			}
-		})
-	}
-}
-
 func TestDeleteEnvVars(t *testing.T) {
 	tests := []struct {
 		name           string
@@ -4191,242 +2878,6 @@ func TestIsVolumeAnEmptyDir(t *testing.T) {
 	}
 }
 
-func Test_GetInputEnvVarsFromStrings(t *testing.T) {
-	tests := []struct {
-		name          string
-		envVars       []string
-		wantedEnvVars []corev1.EnvVar
-		wantErr       bool
-	}{
-		{
-			name:    "Test case 1: with valid two key value pairs",
-			envVars: []string{"key=value", "key1=value1"},
-			wantedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-				{
-					Name:  "key1",
-					Value: "value1",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:    "Test case 2: one env var with missing value",
-			envVars: []string{"key=value", "key1="},
-			wantedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-				{
-					Name:  "key1",
-					Value: "",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:    "Test case 3: one env var with no value and no =",
-			envVars: []string{"key=value", "key1"},
-			wantErr: true,
-		},
-		{
-			name:    "Test case 4: one env var with multiple values",
-			envVars: []string{"key=value", "key1=value1=value2"},
-			wantedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-				{
-					Name:  "key1",
-					Value: "value1=value2",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:    "Test case 5: two env var with same key",
-			envVars: []string{"key=value", "key=value1"},
-			wantErr: true,
-		},
-		{
-			name:    "Test case 6: one env var with base64 encoded value",
-			envVars: []string{"key=value", "key1=SSd2ZSBnb3QgYSBsb3ZlbHkgYnVuY2ggb2YgY29jb251dHMhCg=="},
-			wantedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-				{
-					Name:  "key1",
-					Value: "SSd2ZSBnb3QgYSBsb3ZlbHkgYnVuY2ggb2YgY29jb251dHMhCg==",
-				},
-			},
-			wantErr: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			envVars, err := GetInputEnvVarsFromStrings(tt.envVars)
-
-			if err == nil && !tt.wantErr {
-				if !reflect.DeepEqual(tt.wantedEnvVars, envVars) {
-					t.Errorf("corev1.Env values are not matching with expected values, expected: %v, got %v", tt.wantedEnvVars, envVars)
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
-func TestGetEnvVarsFromDC(t *testing.T) {
-	tests := []struct {
-		name            string
-		dcName          string
-		projectName     string
-		returnedDC      appsv1.DeploymentConfig
-		returnedEnvVars []corev1.EnvVar
-		wantErr         bool
-	}{
-		{
-			name:        "case 1: with valid existing dc and one valid env var pair",
-			dcName:      "nodejs-app",
-			projectName: "project",
-			returnedDC: appsv1.DeploymentConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs-app",
-				},
-				Spec: appsv1.DeploymentConfigSpec{
-					Template: &corev1.PodTemplateSpec{
-						Spec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{
-									Env: []corev1.EnvVar{
-										{
-											Name:  "key",
-											Value: "value",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			returnedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:        "case 2: with valid existing dc and two valid env var pairs",
-			dcName:      "nodejs-app",
-			projectName: "project",
-			returnedDC: appsv1.DeploymentConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "nodejs-app",
-				},
-				Spec: appsv1.DeploymentConfigSpec{
-					Template: &corev1.PodTemplateSpec{
-						Spec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{
-									Env: []corev1.EnvVar{
-										{
-											Name:  "key",
-											Value: "value",
-										},
-										{
-											Name:  "key-1",
-											Value: "value-1",
-										},
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			returnedEnvVars: []corev1.EnvVar{
-				{
-					Name:  "key",
-					Value: "value",
-				},
-				{
-					Name:  "key-1",
-					Value: "value-1",
-				},
-			},
-			wantErr: false,
-		},
-		{
-			name:        "case 3: with non valid existing dc",
-			dcName:      "nodejs-app",
-			projectName: "project",
-			returnedDC: appsv1.DeploymentConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name: "wildfly-app",
-				},
-				Spec: appsv1.DeploymentConfigSpec{
-					Template: &corev1.PodTemplateSpec{
-						Spec: corev1.PodSpec{
-							Containers: []corev1.Container{
-								{
-									Env: []corev1.EnvVar{},
-								},
-							},
-						},
-					},
-				},
-			},
-			returnedEnvVars: []corev1.EnvVar{},
-			wantErr:         false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient, fakeClientSet := FakeNew()
-
-			fakeClientSet.AppsClientset.PrependReactor("get", "deploymentconfigs", func(action ktesting.Action) (handled bool, ret runtime.Object, err error) {
-				dcName := action.(ktesting.GetAction).GetName()
-				if dcName != tt.dcName {
-					return true, nil, fmt.Errorf("get dc called with different name, expected: %s, got %s", tt.dcName, dcName)
-				}
-				return true, &tt.returnedDC, nil
-			})
-
-			envVars, err := fakeClient.GetEnvVarsFromDC(tt.dcName)
-
-			if err == nil && !tt.wantErr {
-				// Check for validating actions performed
-				if len(fakeClientSet.AppsClientset.Actions()) != 1 {
-					t.Errorf("expected 1 action in GetBuildConfigFromName got: %v", fakeClientSet.AppsClientset.Actions())
-				}
-
-				if !reflect.DeepEqual(tt.returnedEnvVars, envVars) {
-					t.Errorf("env vars are not matching with expected values, expected: %s, got %s", tt.returnedEnvVars, envVars)
-				}
-			} else if err == nil && tt.wantErr {
-				t.Error("error was expected, but no error was returned")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, no error was expected, but got unexpected error: %s", err)
-			}
-		})
-	}
-}
-
 func Test_updateEnvVar(t *testing.T) {
 	type args struct {
 		dc           *appsv1.DeploymentConfig
@@ -4644,189 +3095,6 @@ func Test_findContainer(t *testing.T) {
 	}
 }
 
-func TestCreateBuildConfig(t *testing.T) {
-	type args struct {
-		commonObjectMeta metav1.ObjectMeta
-		namespace        string
-		builderImage     string
-		gitURL           string
-		gitRef           string
-		envVars          []corev1.EnvVar
-	}
-	tests := []struct {
-		name    string
-		args    args
-		wantErr bool
-		actions int
-	}{
-		{
-			name: "Case 1 - Generate and create the BuildConfig",
-			args: args{
-				builderImage: "ruby:latest",
-				namespace:    "testing",
-				gitURL:       "https://github.com/openshift/ruby",
-				gitRef:       "master",
-				commonObjectMeta: metav1.ObjectMeta{
-					Name: "ruby",
-					Labels: map[string]string{
-						"app":                        "apptmp",
-						"app.kubernetes.io/instance": "ruby",
-						"app.kubernetes.io/name":     "ruby",
-						"app.kubernetes.io/part-of":  "apptmp",
-					},
-					Annotations: map[string]string{
-						"app.openshift.io/vcs-uri":                "https://github.com/openshift/ruby",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-				envVars: []corev1.EnvVar{
-					{
-						Name:  "key",
-						Value: "value",
-					},
-					{
-						Name:  "key1",
-						Value: "value1",
-					},
-				},
-			},
-			wantErr: false,
-			actions: 1,
-		},
-		{
-			name: "Case 2 - Generate and create the BuildConfig but fail with unable to find image name",
-			args: args{
-				builderImage: "fakeimagename:notlatest",
-				namespace:    "testing",
-				gitURL:       "https://github.com/openshift/ruby",
-				commonObjectMeta: metav1.ObjectMeta{
-					Name: "ruby",
-					Labels: map[string]string{
-						"app":                        "apptmp",
-						"app.kubernetes.io/instance": "ruby",
-						"app.kubernetes.io/name":     "ruby",
-						"app.kubernetes.io/part-of":  "apptmp",
-					},
-					Annotations: map[string]string{
-						"app.openshift.io/vcs-uri":                "https://github.com/openshift/ruby",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-				envVars: []corev1.EnvVar{
-					{
-						Name:  "key",
-						Value: "value",
-					},
-					{
-						Name:  "key1",
-						Value: "value1",
-					},
-				},
-			},
-			wantErr: true,
-			actions: 1,
-		},
-		{
-			name: "Case 3 - Generate and create the BuildConfig but fail with unable to parse image name",
-			args: args{
-				builderImage: "::",
-				namespace:    "testing",
-				gitURL:       "https://github.com/openshift/ruby",
-				commonObjectMeta: metav1.ObjectMeta{
-					Name: "ruby",
-					Labels: map[string]string{
-						"app":                        "apptmp",
-						"app.kubernetes.io/instance": "ruby",
-						"app.kubernetes.io/name":     "ruby",
-						"app.kubernetes.io/part-of":  "apptmp",
-					},
-					Annotations: map[string]string{
-						"app.openshift.io/vcs-uri":                "https://github.com/openshift/ruby",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-				envVars: []corev1.EnvVar{
-					{
-						Name:  "key",
-						Value: "value",
-					},
-					{
-						Name:  "key1",
-						Value: "value1",
-					},
-				},
-			},
-			wantErr: true,
-			actions: 1,
-		},
-		{
-			name: "Case 4 - Generate and create the BuildConfig and pass in no envVars",
-			args: args{
-				builderImage: "ruby:latest",
-				namespace:    "testing",
-				gitURL:       "https://github.com/openshift/ruby",
-				gitRef:       "develop",
-				commonObjectMeta: metav1.ObjectMeta{
-					Name: "ruby",
-					Labels: map[string]string{
-						"app":                        "apptmp",
-						"app.kubernetes.io/instance": "ruby",
-						"app.kubernetes.io/name":     "ruby",
-						"app.kubernetes.io/part-of":  "apptmp",
-					},
-					Annotations: map[string]string{
-						"app.openshift.io/vcs-uri":                "https://github.com/openshift/ruby",
-						"app.kubernetes.io/component-source-type": "git",
-					},
-				},
-				envVars: []corev1.EnvVar{},
-			},
-			wantErr: false,
-			actions: 1,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fakeClient, fakeClientSet := FakeNew()
-
-			fakeClientSet.ImageClientset.PrependReactor("get", "imagestreams", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, fakeImageStream(tt.args.commonObjectMeta.Name, tt.args.commonObjectMeta.Namespace, []string{"latest"}), nil
-			})
-
-			// Run function CreateBuildConfig
-			bc, err := fakeClient.CreateBuildConfig(tt.args.commonObjectMeta, tt.args.builderImage, tt.args.gitURL, tt.args.gitRef, tt.args.envVars)
-
-			if err == nil && !tt.wantErr {
-				// Check to see how many actions are being ran
-				if (len(fakeClientSet.ImageClientset.Actions()) != tt.actions) && !tt.wantErr {
-					t.Errorf("expected %v action(s) in CreateBuildConfig got %v: %v", tt.actions, len(fakeClientSet.ImageClientset.Actions()), fakeClientSet.ImageClientset.Actions())
-				}
-
-				// Check to see that names match
-				if bc.ObjectMeta.Name != tt.args.commonObjectMeta.Name {
-					t.Errorf("Expected buildConfig name %s, got '%s'", tt.args.commonObjectMeta.Name, bc.ObjectMeta.Name)
-				}
-
-				// Check to see that labels match
-				if !reflect.DeepEqual(tt.args.commonObjectMeta.Labels, bc.ObjectMeta.Labels) {
-					t.Errorf("Expected equal labels, got %+v, expected %+v", tt.args.commonObjectMeta.Labels, bc.ObjectMeta.Labels)
-				}
-
-				// Check to see that annotations match
-				if !reflect.DeepEqual(tt.args.commonObjectMeta.Annotations, bc.ObjectMeta.Annotations) {
-					t.Errorf("Expected equal annotations, got %+v, expected %+v", tt.args.commonObjectMeta.Annotations, bc.ObjectMeta.Annotations)
-				}
-
-			} else if err == nil && tt.wantErr {
-				t.Error("test failed, expected: false, got true")
-			} else if err != nil && !tt.wantErr {
-				t.Errorf("test failed, expected: no error, got error: %s", err.Error())
-			}
-
-		})
-	}
-}
-
 func TestGetServiceClassesByCategory(t *testing.T) {
 	t.Run("GetServiceClassesByCategory should work", func(t *testing.T) {
 		client, fakeClientSet := FakeNew()
@@ -4891,71 +3159,6 @@ func TestGetMatchingPlans(t *testing.T) {
 			t.Errorf("test failed, expected %v, got %v", expected, plans)
 		}
 	})
-}
-
-func TestGetPortsFromBuilderImage(t *testing.T) {
-
-	type args struct {
-		componentType string
-	}
-	tests := []struct {
-		name           string
-		imageNamespace string
-		args           args
-		want           []string
-		wantErr        bool
-	}{
-		{
-			name:           "component type: nodejs",
-			imageNamespace: "openshift",
-			args:           args{componentType: "nodejs"},
-			want:           []string{"8080/TCP"},
-			wantErr:        false,
-		},
-		{
-			name:           "component type: php",
-			imageNamespace: "openshift",
-			args:           args{componentType: "php"},
-			want:           []string{"8080/TCP", "8443/TCP"},
-			wantErr:        false,
-		},
-		{
-			name:           "component type: is empty",
-			imageNamespace: "openshift",
-			args:           args{componentType: ""},
-			want:           []string{},
-			wantErr:        true,
-		},
-		{
-			name:           "component type: is invalid",
-			imageNamespace: "openshift",
-			args:           args{componentType: "abc"},
-			want:           []string{},
-			wantErr:        true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			fkclient, fkclientset := FakeNew()
-			fkclient.Namespace = "testing"
-			// Fake getting image stream
-			fkclientset.ImageClientset.PrependReactor("get", "imagestreams", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, fakeImageStream(tt.args.componentType, tt.imageNamespace, []string{"latest"}), nil
-			})
-
-			fkclientset.ImageClientset.PrependReactor("get", "imagestreamimages", func(action ktesting.Action) (bool, runtime.Object, error) {
-				return true, fakeImageStreamImage(tt.args.componentType, tt.want, ""), nil
-			})
-			got, err := fkclient.GetPortsFromBuilderImage(tt.args.componentType)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("Client.GetPortsFromBuilderImage() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if !sliceEqual(got, tt.want) {
-				t.Errorf("Client.GetPortsFromBuilderImage() = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
 
 // sliceEqual checks equality of two slices irrespective of the element ordering
